backend/requirements.txt
# pip install -r requirements.txt

asgiref==3.7.2
channels==4.0.0
Django==4.2.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2023.4
sqlparse==0.4.4
tzdata==2023.4